requirements.txt
streamlit
transformers
torch
scikit-learn
pandas
joblib
markdown